Moving the recoil folder into the src/ folder from src/components/
