+++ src/components/grid/Tile.tsx
@@ -4,7 +4,7 @@ import { TileId } from "@/types";
 import { GridItem, GridItemProps, useColorMode } from "@chakra-ui/react";
 import React from "react";
 import { useRecoilValue } from "recoil";
-import { userSettingState } from "@/components/recoil/UserSettingsAtom";
+import { userSettingState } from "@/recoil/UserSettingsAtom";
 import dynamic from "next/dynamic";
 const TileContainer = dynamic(() => import("@/components/grid/TileContainer"));
